+++ packages/dev/compiler/rollup/routeModules.ts
@@ -1,7 +1,7 @@
 import fs from "fs";
 import type { Plugin } from "rollup";
+import { BuildTarget } from "@remix-run/core";
 
-import { BuildTarget } from "../build";
 import { getRemixConfig } from "./remixConfig";
 
 /**